requirements.txt
fastapi
uvicorn[standard]
groq
requests
networkx
matplotlib
python-dotenv
PyGithub
